fixup! fixup! feat(automation): new world doc pr creation
