+++ internal/docs/mdx.go
@@ -10,36 +10,27 @@ import (
 	"github.com/spf13/cobra"
 )
 
-func GenMdxFile(cmd *cobra.Command, dir string) error {
-	for _, c := range cmd.Commands() {
-		if c.Hidden || c.Name() == "help" {
-			continue
-		}
-		command := newCommand(c)
-		command.Slug = strings.ReplaceAll(command.Name, " ", "-")
-		filename := filepath.Join(dir, fmt.Sprintf("%s.mdx", command.Slug))
+func GenMdxTree(cmd *cobra.Command, dir string) error {
+	tpl, err := template.New("mdx.tpl").Funcs(template.FuncMap{
+		"getExamples": func(cmd Command) []Example {
+			return cmd.ExamplesList()
+		},
+	}).ParseFiles("internal/docs/mdx.tpl")
+	if err != nil {
+		return err
+	}
+
+	commands := getCommands(cmd)
+
+	for _, c := range commands {
+		c.Slug = strings.ReplaceAll(c.Name, " ", "-")
+		filename := filepath.Join(dir, fmt.Sprintf("%s.mdx", c.Slug))
 		file, err := os.Create(filename)
 		if err != nil {
 			return err
 		}
 
-		tpl, err := template.ParseFiles("internal/docs/mdx.tpl")
-		if err != nil {
-			return err
-		}
-
-		if c.HasAvailableSubCommands() {
-			for _, s := range c.Commands() {
-				sub := newCommand(s)
-				if s.HasAvailableSubCommands() {
-					for _, sus := range s.Commands() {
-						sub.SubCommands = append(sub.SubCommands, newCommand(sus))
-					}
-				}
-				command.SubCommands = append(command.SubCommands, sub)
-			}
-		}
-		err = tpl.Execute(file, command)
+		err = tpl.Execute(file, c)
 		if err != nil {
 			return err
 		}
